requirements.txt
torch-ema
ninja
trimesh
opencv-python
tensorboardX
numpy 
pandas
tqdm
matplotlib
PyMCubes
rich
dearpygui
packaging
scipy
scikit-learn

face_alignment
python_speech_features
numba
resampy
#pyaudio
soundfile==0.12.1
einops
configargparse

lpips==0.1.3
imageio-ffmpeg

transformers==4.46.2
edge_tts
flask
flask_sockets
opencv-python-headless
aiortc
aiohttp_cors

ffmpeg-python
omegaconf
diffusers
accelerate

librosa
openai
websocket-client
websockets==12.0
#gradio_client
#azure-cognitiveservices-speech

typeguard==2.13.3
onnxruntime-gpu

# Chat History
aiosqlite>=0.20.0

# RAG Knowledge Base
chromadb>=0.5.0
langchain-text-splitters>=0.3.0
langchain-core>=0.3.0
pypdf>=4.0.0
python-docx>=1.1.0
PyYAML>=6.0.1
